+++ CHANGELOG.md
@@ -1,11 +1,16 @@
-## 0.11.1 - 2024-02-01
+## 0.12 0 - 2924-02-01
+* Add Onchain endpoints
+* Add Examples
+* Add custom User-Agent
+
+## 0.11.1 - 2024-01-26
 * Add Documentations
 * Add missing parameters in simple/price
 
-## 0.11 - 2024-02-01
+## 0.11 - 2024-01-26
 * Fix module not found error for ProClient
 
-## 0.10 - 2024-02-01
+## 0.10 - 2024-01-26
 * First version of the client